+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 
-## 0.20.1
+## 0.21.0
 
-BUG FIXES:
+*June 20th, 2018*
 
+BREAKING CHANGES
+
+- [config] Change default ports from 4665X to 2665X
+
+BUG FIXES
+
+- [consensus] Fix #1754 where we don't make blocks when `create_empty_blocks=false`
+- [mempool] Fix #1761 where we don't process txs if `cache_size=0`
 - [rpc] fix memory leak in Websocket (when using `/subscribe` method)
 
 ## 0.20.0